Combine ssh_login and ssh_login_pubkey modules
